+++ CHANGELOG.md
@@ -36,6 +36,7 @@ contents to a new release section (e.g. _[0.x]_).
   transforming them to equations first and taking the complement of the solutions (PLUT-639)
 - SDK: added support for rendering the set subtraction operator and rendering solutions with holes, e.g. `R \ {1, 2}`
   as `x != 1 and x != 2` (PLUT-639)
+- SDK: exported another call signature for `getInnerSteps`, exported `substitute` function (SP-82)
 
 ### Fixed